cli(refactor): add changefeed cmds
