+++ pkg/cmd/cli/cli_changefeed_helper.go
@@ -152,3 +152,46 @@ func sendOwnerChangefeedQuery(ctx context.Context, etcdClient *kv.CDCEtcdClient,
 
 	return string(body), nil
 }
+
+// sendOwnerAdminChangeQuery sends owner admin query request.
+func sendOwnerAdminChangeQuery(ctx context.Context, etcdClient *kv.CDCEtcdClient, job model.AdminJob, credential *security.Credential) error {
+	owner, err := getOwnerCapture(ctx, etcdClient)
+	if err != nil {
+		return err
+	}
+
+	scheme := util.HTTP
+	if credential.IsTLSEnabled() {
+		scheme = util.HTTPS
+	}
+
+	url := fmt.Sprintf("%s://%s/capture/owner/admin", scheme, owner.AdvertiseAddr)
+	httpClient, err := httputil.NewClient(credential)
+	if err != nil {
+		return err
+	}
+
+	forceRemoveOpt := "false"
+	if job.Opts != nil && job.Opts.ForceRemove {
+		forceRemoveOpt = "true"
+	}
+
+	resp, err := httpClient.PostForm(url, map[string][]string{
+		cdc.APIOpVarAdminJob:           {fmt.Sprint(int(job.Type))},
+		cdc.APIOpVarChangefeedID:       {job.CfID},
+		cdc.APIOpForceRemoveChangefeed: {forceRemoveOpt},
+	})
+	if err != nil {
+		return err
+	}
+
+	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
+		body, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			return errors.BadRequestf("admin changefeed failed")
+		}
+		return errors.BadRequestf("%s", string(body))
+	}
+
+	return nil
+}
