+++ pkg/cmd/cli/cli_changefeed_pause.go
@@ -87,5 +87,6 @@ func newCmdPauseChangefeed(f factory.Factory) *cobra.Command {
 	}
 
 	o.addFlags(command)
+
 	return command
 }
